Extract AssetsInstaller and PackageInstallerWizard
