+++ Editor/Packages.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: faa11f3fd65ae4545916bc9d294b58e4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
