+++ Editor/Packages/AssetsInstaller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e009120ab427f464f90e5e814b29ce7f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
