+++ Editor/Packages/PackageInstallerWizard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 274b6ae4a7cac0243bf75f748f96ad81
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
